requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.2
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-redis==5.2.0
django-rest-auth==0.9.5
django-timezone-field==5.0
djangorestframework==3.14.0
drf-jwt==1.19.2
executing==1.1.1
flake8==5.0.4
Flask==2.2.2
importlib-metadata==5.0.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyee==9.0.4
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
six==1.16.0
slack-sdk==3.19.1
slackeventsapi==3.0.1
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.5
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
